better criterion for feature match

+++ tracker.hpp
@@ -52,7 +52,7 @@ public:
     Tracker();
     void changeParam(int mnf, float ql, int md, int bs, float bv, float nmt, float rt, int bds, int ng, int pg, int bsg, int bvg, float mtg, float rtg);
     void setTarget(const Mat& frame);
-    int featureMatch(const Mat& frame, Mat& homography, bool showImg=false, string windowName="No Name");
+    double featureMatch(const Mat& frame, Mat& homography, bool showImg=false, string windowName="No Name");
     void gridInline(const Mat &matchedFrameCropped, vector<Point2f> *inline_matched, Mat &homo);
     Mat pixelMatch(const Mat& recMatchedFrame);
